+++ sample-apps/react/cookbook-participant-list/src/App.tsx
@@ -4,6 +4,7 @@ import { CallSetup } from './CallSetup';
 import { useEffect, useState } from 'react';
 import {
   StreamMeeting,
+  StreamTheme,
   StreamVideo,
   useCreateStreamVideoClient,
 } from '@stream-io/video-react-sdk';
@@ -39,7 +40,7 @@ const App = () => {
   }, [callId]);
 
   return (
-    <main className="main-container str-video">
+    <StreamTheme as="main" className="main-container">
       <ThemeProvider theme={theme}>
         {!callId && <CallSetup onJoin={setCallId} />}
         {callId && (
@@ -48,13 +49,14 @@ const App = () => {
               callId={callId}
               callType="default"
               data={{ create: true }}
+              autoJoin
             >
               <SpeakerView />
             </StreamMeeting>
           </StreamVideo>
         )}
       </ThemeProvider>
-    </main>
+    </StreamTheme>
   );
 };