+++ sample-apps/react/react-video-demo/src/components/Footer/Footer.tsx
@@ -60,14 +60,15 @@ export const Footer: FC<Props> = ({
     toggleParticipants,
   } = usePanelContext();
 
-  const [showSettingsPanel, setShowSettingsPanel] = useState<boolean>(true);
+  const [showSettingsPanel, setShowSettingsPanel] = useState<boolean>(false);
   const [showReactionsPanel] = useState<boolean>(false);
 
   useEffect(() => {
-    if (isVisible && showSettingsPanel) {
+    console.log('MODAL VISIBLE', isVisible);
+    if (isVisible) {
       setShowSettingsPanel(false);
     }
-  }, [showSettingsPanel, isVisible]);
+  }, [isVisible]);
 
   const recordClassNames = classnames(styles.record, {
     [styles.recording]: isRecording,