+++ packages/client/CHANGELOG.md
@@ -2,6 +2,104 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.0.1-alpha.137](https://github.com/GetStream/stream-video-js/compare/client0.0.1-alpha.136...client0.0.1-alpha.137) (2023-05-12)
+
+
+
+## [0.0.1-alpha.136](https://github.com/GetStream/stream-video-js/compare/client0.0.1-alpha.135...client0.0.1-alpha.136) (2023-05-12)
+
+
+### Features
+
+* Livestream sample app ([#489](https://github.com/GetStream/stream-video-js/issues/489)) ([bf6e6e5](https://github.com/GetStream/stream-video-js/commit/bf6e6e54dab884828ca08208f25b1285cf3f1944))
+
+
+
+## [0.0.1-alpha.135](https://github.com/GetStream/stream-video-js/compare/client0.0.1-alpha.134...client0.0.1-alpha.135) (2023-05-12)
+
+
+### Bug Fixes
+
+* disable sdp munging ([#491](https://github.com/GetStream/stream-video-js/issues/491)) ([5bca195](https://github.com/GetStream/stream-video-js/commit/5bca195dbb99efb77c5d7106a251e88d9a045498)), closes [#395](https://github.com/GetStream/stream-video-js/issues/395)
+
+
+### Features
+
+* enable codec preferring through sdp munging ([#395](https://github.com/GetStream/stream-video-js/issues/395)) ([a5db5a8](https://github.com/GetStream/stream-video-js/commit/a5db5a831ccfc833ecfc29e22540be0f0121287d))
+
+
+
+## [0.0.1-alpha.134](https://github.com/GetStream/stream-video-js/compare/client0.0.1-alpha.133...client0.0.1-alpha.134) (2023-05-11)
+
+
+### Features
+
+* server-side participant counting ([#481](https://github.com/GetStream/stream-video-js/issues/481)) ([3410c25](https://github.com/GetStream/stream-video-js/commit/3410c25ec1449f4d39b44080ad64238b38446612))
+
+
+
+## [0.0.1-alpha.133](https://github.com/GetStream/stream-video-js/compare/client0.0.1-alpha.132...client0.0.1-alpha.133) (2023-05-11)
+
+
+
+## [0.0.1-alpha.132](https://github.com/GetStream/stream-video-js/compare/client0.0.1-alpha.131...client0.0.1-alpha.132) (2023-05-11)
+
+
+
+## [0.0.1-alpha.131](https://github.com/GetStream/stream-video-js/compare/client0.0.1-alpha.130...client0.0.1-alpha.131) (2023-05-10)
+
+
+### Features
+
+* call broadcasting events and apis ([#478](https://github.com/GetStream/stream-video-js/issues/478)) ([8aa4fcb](https://github.com/GetStream/stream-video-js/commit/8aa4fcbbe00b43d4d9f977daa2fd297436e83b51))
+
+
+
+## [0.0.1-alpha.130](https://github.com/GetStream/stream-video-js/compare/client0.0.1-alpha.129...client0.0.1-alpha.130) (2023-05-10)
+
+
+
+## [0.0.1-alpha.129](https://github.com/GetStream/stream-video-js/compare/client0.0.1-alpha.128...client0.0.1-alpha.129) (2023-05-09)
+
+
+### Bug Fixes
+
+* bugfixes in event handling and blocking users ([#476](https://github.com/GetStream/stream-video-js/issues/476)) ([5e9a4fb](https://github.com/GetStream/stream-video-js/commit/5e9a4fb2819cc087f3e095b93488fa9c4df018a4))
+
+
+
+## [0.0.1-alpha.128](https://github.com/GetStream/stream-video-js/compare/client0.0.1-alpha.127...client0.0.1-alpha.128) (2023-05-09)
+
+
+### Reverts
+
+* Revert "chore: Avoid non-atomic pushes resulting in invalid version tags" ([24be283](https://github.com/GetStream/stream-video-js/commit/24be28300a5f5d452338457d60b1e34682027be6))
+
+
+
+## [0.0.1-alpha.127](https://github.com/GetStream/stream-video-js/compare/client0.0.1-alpha.126...client0.0.1-alpha.127) (2023-05-09)
+
+
+### Bug Fixes
+
+* **permissions:** add shortcut APIs ([#475](https://github.com/GetStream/stream-video-js/issues/475)) ([466614c](https://github.com/GetStream/stream-video-js/commit/466614ccf7f15cb16fffc97d5aa05318d2adf7e7))
+* preserve last used deviceId when muting ([#470](https://github.com/GetStream/stream-video-js/issues/470)) ([b4ed2d0](https://github.com/GetStream/stream-video-js/commit/b4ed2d0cce1456a748f070bc01a68df574683862))
+
+
+
+## [0.0.1-alpha.126](https://github.com/GetStream/stream-video-js/compare/client0.0.1-alpha.125...client0.0.1-alpha.126) (2023-05-08)
+
+
+### Features
+
+* **latency:** breadth-first time-boxed latency measurement ([#467](https://github.com/GetStream/stream-video-js/issues/467)) ([6d66003](https://github.com/GetStream/stream-video-js/commit/6d660032064667586902a9b410ee6dabfcc6b7ba))
+
+
+
+## [0.0.1-alpha.125](https://github.com/GetStream/stream-video-js/compare/client0.0.1-alpha.124...client0.0.1-alpha.125) (2023-05-08)
+
+
+
 ## [0.0.1-alpha.124](https://github.com/GetStream/stream-video-js/compare/client0.0.1-alpha.123...client0.0.1-alpha.124) (2023-05-05)